requirements.txt
streamlit
openai>=1.0.0
langchain
langchain-openai
faiss-cpu
sentence-transformers
pypdf2
pandas
python-dotenv
tiktoken
openpyxl
